+++ src/core/data_basic.jl
@@ -17,6 +17,30 @@ users requiring any of the features listed above for their analysis should use
 the non-basic PowerModels routines.
 """
 function make_basic_network(data::Dict{String,<:Any})
+    # These initial checks are redundant with the checks in make_basic_network!
+    # We keep them here so they run _before_ we create a deepcopy of the data
+    # The checks are fast so this redundancy has little performance impact
+    if _IM.ismultiinfrastructure(data)
+        Memento.error(_LOGGER, "make_basic_network does not support multiinfrastructure data")
+    end
+    
+    if _IM.ismultinetwork(data)
+        Memento.error(_LOGGER, "make_basic_network does not support multinetwork data")
+    end
+
+    data = deepcopy(data)
+
+    make_basic_network!(data)
+
+    return data
+end
+
+"""
+given a powermodels data dict, modifies it in-place to conform to basic network model requirements.
+
+See [`make_basic_network`](@ref) for more information.
+"""
+function make_basic_network!(data::Dict{String,<:Any})
     if _IM.ismultiinfrastructure(data)
         Memento.error(_LOGGER, "make_basic_network does not support multiinfrastructure data")
     end
@@ -24,9 +48,6 @@ function make_basic_network(data::Dict{String,<:Any})
     if _IM.ismultinetwork(data)
         Memento.error(_LOGGER, "make_basic_network does not support multinetwork data")
     end
-
-    # make a copy of data so that modifications do not change the input data
-    data = deepcopy(data)
 
     # TODO transform PWL costs into linear costs
     for (i,gen) in data["gen"]
@@ -110,7 +131,7 @@ function make_basic_network(data::Dict{String,<:Any})
     # re-number non-bus component ids
     for comp_key in keys(pm_component_status)
         if comp_key != "bus"
-            data[comp_key] = _renumber_components(data[comp_key])
+            data[comp_key] = _renumber_components!(data[comp_key])
         end
     end
 
@@ -148,18 +169,18 @@ end
 given a component dict returns a new dict where components have been renumbered
 from 1-to-n ordered by the increasing values of the orginal component id.
 """
-function _renumber_components(comp_dict::Dict{String,<:Any})
-    renumbered_dict = Dict{String,Any}()
+function _renumber_components!(comp_dict::Dict{String,<:Any})
+    comp_ordered = sort([(comp["index"], comp) for (i, comp) in comp_dict], by=(x) -> x[1])
 
-    comp_ordered = sort([comp for (i,comp) in comp_dict], by=(x) -> x["index"])
-
-    for (i,comp) in enumerate(comp_ordered)
-        comp = deepcopy(comp)
-        comp["index"] = i
-        renumbered_dict["$i"] = comp
+    # Delete existing keys
+    empty!(comp_dict)
+    # Update component indices and re-build the dict keys
+    for (i_new, (i_old, comp)) in enumerate(comp_ordered)
+        comp["index"] = i_new
+        comp_dict["$(i_new)"] = comp
     end
 
-    return renumbered_dict
+    return comp_dict
 end